+++ src/test-jmh/groovy/org/xtraktor/RawPointDynamic.groovy
@@ -41,7 +41,7 @@ class RawPointDynamic {
         return LongStream.rangeClosed(minIndex, maxIndex)
                 .parallel()
                 .mapToObj({
-            long pointTime = config.timeMin + config.timeDelta * (it as Long)
+            long pointTime = config.timeMin + config.timeDelta * it
             def pointRatio = (pointTime - timestamp) / (nextPoint.timestamp - timestamp)
 
             double pointLon = ((longitude + (nextPoint.longitude - longitude) * pointRatio) as Double)
